requirements.txt
asgiref==3.7.2
black==23.3.0
click==8.1.3
Django==4.2.2
django-pytest==0.2.0
djangorestframework==3.14.0
freezegun==1.2.2
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pluggy==1.2.0
psycopg2-binary==2.9.6
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
